#137 Constructors added in VFA classes for easier testing. Removed test files because they are no longer needed. Now all tests pass on x64

+++ RLSimion/featuremap-discrete.cpp
@@ -5,7 +5,7 @@
 
 DiscreteFeatureMap::DiscreteFeatureMap(Descriptor& stateDescriptor, vector<size_t> stateVariableIds, Descriptor& actionDescriptor, vector<size_t> actionVariableIds, size_t numFeaturesPerVariable)
 {
-	m_numFeaturesPerVariable.set(numFeaturesPerVariable);
+	m_numFeaturesPerVariable.set((int)numFeaturesPerVariable);
 
 	//create STATE_VARIABLE objects
 	for each (size_t varId in stateVariableIds)
